github_service/requirements.txt
pulsar-client==2.3.1